chore: improve stability
- change app loading timeout
- change click in object class and button class
- chore: remove unnecessary wait
- add attempts to return the password view
- increase timeout to let the balance show up
- remove waiting functions where possible (caused recursion)
- divide test to multiple to test if the issue is parametrization
