+++ configs/timeouts.py
@@ -3,5 +3,5 @@
 UI_LOAD_TIMEOUT_SEC = 5
 UI_LOAD_TIMEOUT_MSEC = UI_LOAD_TIMEOUT_SEC * 1000
 PROCESS_TIMEOUT_SEC = 10
-APP_LOAD_TIMEOUT_MSEC = 60000
+APP_LOAD_TIMEOUT_MSEC = 90000
 MESSAGING_TIMEOUT_SEC = 60
