+++ tests/settings/settings_wallet/test_wallet_settings_account_order.py
@@ -25,7 +25,7 @@ def test_change_account_order_by_drag_and_drop(main_screen: MainWindow, user_acc
                                                second_color: str, second_emoji: str, second_acc_emoji: str):
     with step('Create watch-only wallet account'):
         wallet = main_screen.left_panel.open_wallet()
-        SigningPhrasePopup().wait_until_appears().confirm_phrase()
+        SigningPhrasePopup().confirm_phrase()
         account_popup = wallet.left_panel.open_add_account_popup()
         account_popup.set_name(name).set_emoji(emoji).set_color(color).set_origin_watched_address(address).save()
         account_popup.wait_until_hidden()
@@ -33,7 +33,7 @@ def test_change_account_order_by_drag_and_drop(main_screen: MainWindow, user_acc
     with step('Create generated wallet account'):
         account_popup = wallet.left_panel.open_add_account_popup()
         account_popup.set_name(second_name).set_emoji(second_emoji).set_color(second_color).save()
-        AuthenticatePopup().wait_until_appears().authenticate(user_account.password)
+        AuthenticatePopup().authenticate(user_account.password)
         account_popup.wait_until_hidden()
 
     with step('Verify accounts in wallet settings'):
