+++ tests/wallet_main_screen/test_wlt_main_add_acc_manage_watched_address.py
@@ -23,7 +23,7 @@ def test_wallet_add_acc_add_watched_address(
         name: str):
     with step('Add watched address with plus action button'):
         wallet = main_screen.left_panel.open_wallet()
-        SigningPhrasePopup().wait_until_appears().confirm_phrase()
+        SigningPhrasePopup().confirm_phrase()
         account_popup = wallet.left_panel.open_add_account_popup()
         account_popup.set_name(name).set_emoji(emoji).set_color(color).set_origin_watched_address(address).save()
         account_popup.wait_until_hidden()
